rename interface to interface_old to facilitate new code
